add : 로그인 시스템 구현

+++ Quest10/client-server/public/notepad.js
@@ -5,6 +5,7 @@ import Editor from "./components/Editor.js";
 import Tab from "./components/Tab.js";
 import Files from "./components/Files.js";
 import HandleApi from "./utils/HandleAPI.js";
+import Login from "./components/Login.js";
 class Notepad {
   #isLoading = false;
   #tabStorage;
@@ -15,6 +16,7 @@ class Notepad {
   #tab;
   #files;
   #button;
+  #login = new Login();
   constructor({
     editorTarget,
     tabTarget,
@@ -32,6 +34,7 @@ class Notepad {
     this.#button.createButton("new");
     this.#button.createButton("save");
     this.#button.createButton("save as");
+    this.#button.createButton("logIn");
     this.#setNewEvent();
     this.#setSaveEvent();
     this.#setSaveAsEvent();
@@ -40,10 +43,16 @@ class Notepad {
     this.#setTabDeleteEvent();
     this.#setFileDeleteEvent();
     this.#setFileClickEvent();
+    this.#setLoginEvent();
     this.#init();
   }
   //api
   async #checkTitle() {
+    const auth = await this.#fileStorage.auth();
+    if (!auth) {
+      alert("로그인을 먼저 해주세요");
+      return;
+    }
     const title = prompt("이름을 입력해주세요");
     const checking = await this.#fileStorage.checkOverLap(title);
     if (checking) return true;
@@ -51,6 +60,11 @@ class Notepad {
   }
   //api
   async #saveFile() {
+    const auth = await this.#fileStorage.auth();
+    if (!auth) {
+      alert("로그인을 먼저 해주세요");
+      return;
+    }
     await this.#fileStorage.updateFile(this.#state);
     this.#tabStorage.updateFile(this.#state);
     const { id } = this.#state;
@@ -60,11 +74,17 @@ class Notepad {
   //api
   #setNewEvent() {
     this.#event.setEvent("new", async () => {
+      const auth = await this.#fileStorage.auth();
+      if (!auth) {
+        alert("로그인을 먼저 해주세요");
+        return;
+      }
       try {
         if (this.#isLoading) {
           console.log("아직 로딩중입니다.");
           return;
         }
+
         this.#isLoading = true;
         const nextState = await this.#fileStorage.createFile();
         this.#tabStorage.insertFile(nextState);
@@ -81,6 +101,11 @@ class Notepad {
   //api
   async #setSaveEvent() {
     this.#event.setEvent("save", async () => {
+      const auth = await this.#fileStorage.auth();
+      if (!auth) {
+        alert("로그인을 먼저 해주세요");
+        return;
+      }
       if (this.#isLoading) return;
       this.#isLoading = true;
       try {
@@ -99,6 +124,11 @@ class Notepad {
   }
   #setSaveAsEvent() {
     this.#event.setEvent("save as", async () => {
+      const auth = await this.#fileStorage.auth();
+      if (!auth) {
+        alert("로그인을 먼저 해주세요");
+        return;
+      }
       if (this.#isLoading) return;
       this.#isLoading = true;
       try {
@@ -115,6 +145,11 @@ class Notepad {
   //api
   #setFileDeleteEvent() {
     this.#event.setEvent("onDeleteFile", async (e) => {
+      const auth = await this.#fileStorage.auth();
+      if (!auth) {
+        alert("로그인을 먼저 해주세요");
+        return;
+      }
       if (!confirm("정말 삭제하시겠습니까?")) return;
       this.#tabStorage.removeFile(e.detail);
       await this.#fileStorage.removeFile(e.detail);
@@ -149,6 +184,11 @@ class Notepad {
   //api
   #setFileClickEvent() {
     this.#event.setEvent("onClickFile", async (e) => {
+      const auth = await this.#fileStorage.auth();
+      if (!auth) {
+        alert("로그인을 먼저 해주세요");
+        return;
+      }
       const data = await this.#fileStorage.getFile(e.detail);
       this.#tabStorage.insertFile(data);
       this.#setState(data);
@@ -160,6 +200,20 @@ class Notepad {
       const nextState = this.#tabStorage.getFile(e.detail);
       this.#setState(nextState);
       this.#editor.setState(this.#state.id);
+    });
+  }
+  #setLoginEvent() {
+    this.#event.setEvent("onlogin", async (e) => {
+      const { id, password } = e.detail;
+      const result = await this.#fileStorage.login(id, password);
+      if (!result) {
+        alert("아아디 혹은 비밀번호가 올바르지 않습니다.");
+        return;
+      }
+      await this.#init();
+      console.log(result);
+      this.#event.dispatch("onKey", result.username);
+      this.#event.dispatch("modalOff");
     });
   }
   #setInputEvent() {
@@ -176,6 +230,9 @@ class Notepad {
     this.#files.setState(id);
   }
   async #init() {
+    const auth = await this.#fileStorage.auth();
+    if (!auth) return;
+    this.#event.dispatch("onKey", auth);
     const initList = await this.#fileStorage.getList();
     const nextState = initList[0] || { id: null };
     this.#setState(nextState);
@@ -192,6 +249,6 @@ new Notepad({
   tabTarget,
   buttonTarget,
   filesTarget,
-  tabStorage: new Storage(window.sessionStorage),
+  tabStorage: new Storage(window.localStorage),
   fileStorage: new HandleApi(),
 });
